program
Visualization and animation help us to learn and understand clearly about any things or any process. In this project I had visualize the rocket launching steps via animation using graphics (Computer Graphics). The main motto of the project is to simulate the real-life launch and make the teaching method easier. It’s very difficult to learn theory-based launch and to solve this problem I had purposed this project. Background sound in the project has made the project more realistic. Through such visual and graphics, it would be fun and easier while teaching.